+++ client/src/main/js/service/ModalService.js
@@ -77,6 +77,30 @@
             });
         };
 
+
+        /**
+         * Abre o modal de visualização da imagem em 'full size'
+         *
+         * @param {Object} imagem Imagem a ser visualizada
+         * @param {Event} $event Evendo do clique.
+         * @param {Boolean} editavel Identifica se a imagem pode ser excluída ou não.
+         * @param {Function} excluirImagemCallback Callback a ser chamado ao apertar o botão de excluir imagem no modal.
+         * @returns {Promise} Promise do modal.
+         */
+        this.verImagem = function (imagem, $event, editavel, excluirImagemCallback) {
+            const options = {
+                templateUrl: 'view/dialog/full-size-image.html',
+                controller: 'ImagemController as imagemCtrl',
+                targetEvent: $event,
+                resolve: {
+                    imagem: () => imagem,
+                    editavel: () => editavel,
+                    excluirImagemCallback : () => excluirImagemCallback
+                }
+            };
+            return self.custom(options);
+        };
+
         /**
          * Abre o modal de visualização para termos do local.
          *